+++ src/modules/localStorage.js
@@ -1,0 +1,13 @@
+function saveTasksToLocalStorage(tasks) {
+  localStorage.setItem('tasks', JSON.stringify(tasks));
+}
+
+function loadTasksFromLocalStorage() {
+  const savedTasks = localStorage.getItem('tasks');
+  if (savedTasks) {
+    return JSON.parse(savedTasks);
+  }
+  return [];
+}
+
+export { saveTasksToLocalStorage, loadTasksFromLocalStorage };
